add LE certificate for the kgb bot relay
The rationale for this tool is to create a mechanism by which
arbitrary "clients" can send notifications to a predetermined set of
channels. While this can be done per host (e.g. running KGB on nagios
or gitlab) there are certain use cases where this is not logical.

For example, in the reboot procedures, we run the jobs from the
servers directly (in the old way) or the sysadmin's machine (the new
way). In that context, it makes no sense to run KGB bots all over the
place and having a central, neutral location actually makes more
sense.

Having an HTTPS certificate is essential for remote operation to be
secure against password hijacking.