Get certs for pin-affected hostnames more often
We set RENEW_DAYS="88" for the certs for
  torproject.org
  blog.torproject.org
  check.torproject.org
  dist.torproject.org
  www.torproject.org
so we should be getting a cert every two days.

The plan is to have one that is signed with the X3 intermediate and that
lasts as long as possible.  Cf.  tpo/tpa/team#40052.

This is combined with a (temporary) deploy hook on the LE host that
errors out if the chain is not through X3.